==> R/03_ms_acs.Rmd <==
---
title: "Microsoft and ACS"
author: "Renata Gerecke"
date: "7/8/2021"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(tidycensus)
library(usmap)
library(janitor)
library(broom)
library(sf)
library(stargazer)
```

Microsoft collects its own data on broadband internet access, reporting the percent of people in a county who are actually accessing "high-speed" internet (25 mbps upload/3 mbps download). These numbers could vary from the FCC estimates for several reasons, including: 

1. The ISPs are reporting that they serve areas that they do not actually support;
2. The ISPs are reporting that they provide speeds that they do not actually achieve;
3. The ISPs serve the area at high speeds but at a price point that is too high for the residents to afford; or
4. The residents aren't interested in high speed internet access.

For this analysis I will be using open-source data from Microsoft that both reports the FCC estimate of availability and Microsoft's observed measure of accessibility. Then, I will use the 5-year ACS population estimates to assess the degree to which demographic characteristics are associated with high-speed internet access.

# Setup

First, I read in two data files: Microsoft's data from November 2019 and the ACS 5-year estimate from 2015-2019.

```{r}
df_ms_19 <- read_csv("../data/ms/broadband_data_2019November.csv",
                  na = "-")

df_ms_20 <- read_csv("../data/ms/broadband_data_2020October.csv",
                     na = "-", skip = 18)

df_acs <- read_rds("../data/acs/acs_5yr_ed.rds")

df_county <- st_read("../data/tl_2019_us_county/tl_2019_us_county.shp")
```

# Munge

The main adjustment that needs to be made to the Micorosft data is the calculation of the difference between FCC's reported broadband availability and Microsoft's observed broadband usage. A difference of .2 would indicate that the FCC has overestimated availability by 20 percentage points; a difference of -.2 would indicate an underestimate of the same magnitude. 

```{r}
df_ms_mut <- bind_rows(df_ms_19, df_ms_20, .id = "year") %>%
    clean_names() %>%
    mutate(delta = broadband_availability_per_fcc - broadband_usage,
           fips = county_id,
           year = ifelse(year == 1, 2019, 2020)) %>%
    pivot_wider(
        names_from = year,
        values_from = c(starts_with("broadband"), delta),
        names_sep = "_"
    )
```

# Map

```{r}
plot_usmap(data = df_ms_mut, values = "broadband_usage_2020", color = "transparent") +
    scale_fill_gradient(label = scales::percent, limits = c(0,1)) + 
    labs(title = "Percent of Residents Using Broadband Internet, October 2020",
         subtitle = "25 mbps download",
         fill = NULL,
         caption = "Source: Microsoft Broadband Data") +
    theme(legend.position = "right")
```


```{r}
plot_usmap(data = df_ms_mut, values = "delta_2019", color = "transparent") +
    scale_fill_gradient2(label = scales::percent, limits = c(-1,1)) + 
    labs(title = "Percentage Point Overestimation of Broadband Internet Access",
         subtitle = "November 2019",
         fill = NULL,
         caption = "Source: Microsoft Broadband Data") + 
    theme(legend.position = "right")

# ggsave("../output/ms_fcc_delta.png")
```

```{r}
plot_usmap(data = df_ms_mut, values = "delta_2020", color = "transparent") +
    scale_fill_gradient2(label = scales::percent, limits = c(-1,1)) + 
    labs(title = "Percentage Point Overestimation of Broadband Internet Access",
         subtitle = "October 2020",
         fill = NULL,
         caption = "Source: Microsoft Broadband Data") + 
    theme(legend.position = "right")
```



# Plot

```{r}

df_merge <- df_ms_mut %>% 
    mutate(fips = as.character(county_id) %>%
               str_pad(5, side = "left", pad = "0")) %>%
    left_join(df_acs, by = "fips") %>%
    left_join(select(st_drop_geometry(df_county), fips = GEOID, area = ALAND), by = "fips") %>%
  mutate(pop_dense = tot_pop / (area / 2589988))

ggplot(df_merge) + 
    aes(x = broadband_availability_per_fcc_2019, y = broadband_usage_2019) + 
    geom_point(alpha = .3) + 
    geom_abline(slope = 1, intercept = 0, color = "red") + 
    scale_x_continuous(label = scales::percent) + 
    scale_y_continuous(label = scales::percent) + 
  labs(
    title = "Availability vs. Usage of Broadband Internet",
    subtitle = "By U.S. County, November 2019",
    caption = "Source: Microsoft Broadband Data",
    x = "Broadband Availability Reported to FCC",
    y = "Broadband Usage Observed by Microsoft"
  ) + 
    theme_classic()
```


```{r}
ggplot(df_merge) + 
    aes(x = broadband_availability_per_fcc_2020, y = broadband_usage_2020) + 
  geom_point(alpha = .3) +
    geom_abline(slope = 1, intercept = 0, color = "red") + 
    scale_x_continuous(label = scales::percent) + 
    scale_y_continuous(label = scales::percent) + 
  labs(
    title = "Availability vs. Usage of Broadband Internet",
    subtitle = "By U.S. County, October 2020",
    caption = "Source: Microsoft Broadband Data",
    x = "Broadband Availability Reported to FCC",
    y = "Broadband Usage Observed by Microsoft"
  ) + 
    theme_classic()
```

```{r}
ggplot(df_merge) + 
    aes(x = log(pop_dense), y = broadband_usage_2020) + 
  geom_point(alpha = .3) +
    # geom_abline(slope = 1, intercept = 0, color = "red") + 
    scale_x_continuous(label = scales::comma) + 
    scale_y_continuous(label = scales::percent) + 
  labs(
    title = "Availability vs. Usage of Broadband Internet",
    subtitle = "By U.S. County, October 2020",
    caption = "Source: Microsoft Broadband Data",
    x = "Population Density (persons per sq.mi.)",
    y = "Broadband Usage Observed by Microsoft"
  ) + 
    theme_classic()
```


# Single Regression

```{r}
df_ms_long <- df_ms_mut %>%
  pivot_longer(
    starts_with("broadband")
  ) %>%
  mutate(year = as.character(parse_number(name)),
         source = case_when(
           str_detect(name, "fcc") ~ "fcc",
           TRUE ~ "ms"
         ),
         delta_2019 = NULL,
         delta_2020 = NULL,
         name = NULL,
         fips = as.character(county_id) %>%
               str_pad(5, side = "left", pad = "0")) %>%
    left_join(df_acs, by = "fips") %>%
  left_join(select(st_drop_geometry(df_county), fips = GEOID, area = ALAND), by = "fips") %>%
  mutate(pop_dense = tot_pop / (area / 2589988))
```

```{r, results = "asis"}
long_lm <- list(
  long_unadj = lm(
    value ~ source * year,
    data = df_ms_long
  ),
  long_adj1 = lm(
    value ~ source * year + race_blacknh + race_asiannh + race_othernh + race_hispanic,
    data = df_ms_long
  ),
  long_adj2 = lm(
    value ~ source * year + race_blacknh + race_asiannh + race_othernh + race_hispanic + male + age_18to64 + age_65up + inc_low + log(pop_dense),
    data = df_ms_long
  )
) 

stargazer(
  long_lm,
  type = "html",
  covariate.labels = c("Source: Microsoft", "Year: 2020", "% Black Non-Hispanic", "% Asian Non-Hispanic", "% Other Non-Hispanic", "% Hispanic", "% Male", "% Age 18-64", "% Age 65+", "% Below Poverty Line", "Log Population Density<br>(persons per sq. mi.)", "Source: Microsoft * Year: 2020"),
  column.labels = c("Unadjusted", "Adjusted by Race", "Adjusted by Demos"), 
  dep.var.labels = "% Residents in County with Broadband Access",
  out = "../output/long_lm.htm"
)


```


```{r}
df_ms_long_20 <- filter(df_ms_long, year == 2020)

long_lm_20 <- list(
  long_unadj = lm(
    value ~ source,
    data = df_ms_long_20
  ),
  long_adj1 = lm(
    value ~ source + race_blacknh + race_asiannh + race_othernh + race_hispanic,
    data = df_ms_long_20
  ),
  long_adj2 = lm(
    value ~ source + race_blacknh + race_asiannh + race_othernh + race_hispanic + male + age_18to64 + age_65up + inc_low + log(pop_dense),
    data = df_ms_long_20
  )
) 

stargazer(
  long_lm_20,
  type = "html",
  covariate.labels = c("Source: Microsoft", "% Black Non-Hispanic", "% Asian Non-Hispanic", "% Other Non-Hispanic", "% Hispanic", "% Male", "% Age 18-64", "% Age 65+", "% Below Poverty Line", "Log Population Density<br>(persons per sq. mi.)"),
  column.labels = c("Unadjusted", "Adjusted by Race", "Adjusted by Demos"), 
  dep.var.labels = "% Residents in County with Broadband Access",
  out = "../output/long_lm_2020.htm"
)
```



# Original Regressions (Unused)

## Match ACS data



**TODO**: Running all these models separately is interesting and says a lot but we can probably estimate the impact of being self-reported vs observed and year-on-year by combining all the data & running a single regression? To try later today.

```{r, eval = FALSE}


all_unadjusted <- list(
  obs_19 = {
    lm(broadband_usage_2019 ~ race_blacknh + race_asiannh + race_othernh + race_hispanic, data = df_merge)
  },
    est_19 = {
    lm(broadband_availability_per_fcc_2019 ~ race_blacknh + race_asiannh + race_othernh + race_hispanic, data = df_merge)
  } , 
  obs_20 = {
    lm(broadband_usage_2020 ~ race_blacknh + race_asiannh + race_othernh + race_hispanic, data = df_merge)
  },
  est_20 = {
    lm(broadband_availability_per_fcc_2020 ~ race_blacknh + race_asiannh + race_othernh + race_hispanic, data = df_merge)
  } 
) 

stargazer(
  all_unadjusted,
  type = "html",
  covariate.labels = c("% Black Non-Hispanic", "% Asian Non-Hispanic", "% Other Non-Hispanic", "% Hispanic"),
  dep.var.labels = c("Observed by MS, 2019", "Estimated by FCC, 2019", "Observed by MS, 2020", "Estimated by FCC, 2020"),
  out = "../output/unadjusted.htm"
)
```


```{r, eval = FALSE}
all_adjusted <- list(
  obs_19 = {
    lm(broadband_usage_2019 ~ race_blacknh + race_asiannh + race_othernh + race_hispanic + male + age_18to64 + age_65up + inc_low, data = df_merge)
  },
  est_19 = {
    lm(broadband_availability_per_fcc_2019 ~ race_blacknh + race_asiannh + race_othernh + race_hispanic + male + age_18to64 + age_65up + inc_low, data = df_merge)
  },
  obs_20 = {
    lm(broadband_usage_2020 ~ race_blacknh + race_asiannh + race_othernh + race_hispanic + male + age_18to64 + age_65up + inc_low, data = df_merge)
  },
  est_20 = {
    lm(broadband_availability_per_fcc_2020 ~ race_blacknh + race_asiannh + race_othernh + race_hispanic + male + age_18to64 + age_65up + inc_low, data = df_merge)
  }
) 
stargazer(
  all_adjusted,
  type = "html",
  covariate.labels = c("% Black Non-Hispanic", "% Asian Non-Hispanic", "% Other Non-Hispanic", "% Hispanic", "% Male", "% Age 18-64", "% Age 65+", "% Below Poverty Line"),
  dep.var.labels = c("Observed by MS, 2019", "Estimated by FCC, 2019", "Observed by MS, 2020", "Estimated by FCC, 2020"),
  out = "../output/adjusted.htm"
)
```
